requirements.txt
Django==3.1.1
django-widget-tweaks
django-countdowntimer-model
psycopg2-binary==2.8.5
gunicorn==20.0.4